pyproject.toml
[tool.poetry]
name = "telegram_bot"
version = "0.1.0"
description = "Telegram Bot"
authors = ["daisukixci <daisuki@tuxtrooper.com>"]

[tool.poetry.dependencies]
python = "^3.9"
croniter = "^1.1.0"
PyYAML = "^6.0"
requests = "^2.27.0"
dokuwiki = "^1.3.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
black = {version = "^22.1.0", allow-prereleases = true}
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
